LayoutTests/fast/forms/ios/inputmode-removing-none-expected.txt

This tests removing inputmode="none" from an input element after the element was focused.
To manually test, detach a hardware keyboard and tap on the text field below. The software keyboard should come up.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



ACTIVATE input with inputmode=none
PASS systemKeyboardRect.height is 0
PASS input.value is "hello,"

Removing inputmode=none
PASS systemKeyboardRect.height > 0 is true
PASS input.value is "hello, world"
PASS successfullyParsed is true

TEST COMPLETE
